chore: split tooltip functionality, units
